requirements.txt
tensorflow
pandas
numpy
streamlit
Pillow